Refactored app.js and PieChart.js for readability
