+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState} from 'react';
 import Chart from "chart.js/auto"
 import { CategoryScale } from "chart.js";
 import PieChart from "./components/PieChart.js";
@@ -8,8 +8,6 @@ import 'leaflet/dist/leaflet.css';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
 import Accordion from 'react-bootstrap/Accordion';
-
-
 
 const states = [
   "Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "Florida", "Georgia", 
@@ -23,55 +21,6 @@ const states = [
 Chart.register(CategoryScale);
 
 function App() {
- // Initializes the charts default state 
-  const [chartData, setChartData] = useState({
-    labels: ['Right-handed', 'Left-handed'],
-    datasets: [
-      {
-        label: "Handedness Distribution",
-        data: [0, 0],
-        backgroundColor: [
-          "rgba(75,192,192,1)",
-          "#ecf0f1"
-        ],
-        borderColor: "black",
-        borderWidth: 2
-      }
-    ]
-  });
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await fetch('https://mp2-backend-production.up.railway.app/api/hand-dominance');
-        const data = await response.json();
-        console.log(data)
-        const rightHanded = data.filter(item => item.choice === 'right').length;
-        const leftHanded = data.filter(item => item.choice === 'left').length;
-
-        setChartData({
-          labels: ['Right-handed', 'Left-handed'],
-          datasets: [
-            {
-              label: "Handedness Distribution",
-              data: [rightHanded, leftHanded],
-              backgroundColor: [
-                "rgba(75,192,192,1)",
-                "#ecf0f1"
-              ],
-              borderColor: "black",
-              borderWidth: 2
-            }
-          ]
-        });
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
-
-    fetchData();
-  }, []);
-
   const [state, setState] = useState('');
   const handleVote = async (choice) => {
     try {
@@ -101,7 +50,7 @@ function App() {
         <Accordion.Header>Pie Chart</Accordion.Header>
         <Accordion.Body>
       
-        <PieChart chartData={chartData} />
+        <PieChart />
       
       </Accordion.Body>
       </Accordion.Item>
@@ -114,7 +63,7 @@ function App() {
         </Accordion.Body>
       </Accordion.Item>
       </Accordion>
-      <div class="selectorArea">
+      <div className="selectorArea">
       <div>
         <Form.Select size="lg" aria-label="Default select example" value={state} onChange={(e) => setState(e.target.value)}>
           <option value="">Select your state</option>
